Set up a Binary Outcome Misclassification jags.model Object for a Given Prior

Set up a Binary Outcome Misclassification jags.model Object for a Given Prior

Usage

jags_picker(
  prior,
  sample_size,
  dim_x,
  dim_z,
  n_cat,
  Ystar,
  X,
  Z,
  beta_prior_parameters,
  gamma_prior_parameters,
  number_MCMC_chains,
  model_file
)

Arguments

prior

A character string specifying the prior distribution for the \(\beta\) and \(\gamma\) parameters. Options are "t", "uniform", "normal", or "dexp" (double Exponential, or Weibull).

sample_size

An integer value specifying the number of observations in the sample.

dim_x

An integer specifying the number of columns of the design matrix of the true outcome mechanism, X.

dim_z

An integer specifying the number of columns of the design matrix of the observation mechanism, Z.

n_cat

An integer specifying the number of categorical values that the true outcome, Y, and the observed outcome, Y* can take.

Ystar

A numeric vector of indicator variables (1, 2) for the observed outcome Y*. The reference category is 2.

X

A numeric design matrix for the true outcome mechanism.

Z

A numeric design matrix for the observation mechanism.

beta_prior_parameters

A numeric list of prior distribution parameters for the \(\beta\) terms. For prior distributions "t", "uniform", "normal", or "dexp", the first element of the list should contain a matrix of location, lower bound, mean, or shape parameters, respectively, for \(\beta\) terms. For prior distributions "t", "uniform", "normal", or "dexp", the second element of the list should contain a matrix of shape, upper bound, standard deviation, or scale parameters, respectively, for \(\beta\) terms. For prior distribution "t", the third element of the list should contain a matrix of the degrees of freedom for \(\beta\) terms. The third list element should be empty for all other prior distributions. All matrices in the list should have dimensions dim_x X n_cat, and all elements in the n_cat column should be set to NA.

gamma_prior_parameters

A numeric list of prior distribution parameters for the \(\gamma\) terms. For prior distributions "t", "uniform", "normal", or "dexp", the first element of the list should contain an array of location, lower bound, mean, or shape parameters, respectively, for \(\gamma\) terms. For prior distributions "t", "uniform", "normal", or "dexp", the second element of the list should contain an array of shape, upper bound, standard deviation, or scale parameters, respectively, for \(\gamma\) terms. For prior distribution "t", the third element of the list should contain an array of the degrees of freedom for \(\gamma\) terms. The third list element should be empty for all other prior distributions. All arrays in the list should have dimensions n_cat X n_cat X dim_z, and all elements in the n_cat row should be set to NA.

number_MCMC_chains

An integer specifying the number of MCMC chains to compute.

model_file

A .BUG file and used for MCMC estimation with rjags.

Value

jags_picker returns a jags.model object for a binay outcome misclassification model. The object includes the specified prior distribution, model, number of chains, and data.

Examples

if (FALSE) {
set.seed(123)
n <- 100
x_mu <- 0
x_sigma <- 1
z_shape <- 1

true_beta <- matrix(c(1, -2), ncol = 1)
true_gamma <- matrix(c(.5, 1, -.5, -1), nrow = 2, byrow = FALSE)

my_data <- COMBO_data(sample_size = n,
                      x_mu = x_mu, x_sigma = x_sigma,
                      z_shape = z_shape,
                      beta = true_beta, gamma = true_gamma)

obs_Y = my_data[["obs_Y"]]
X = my_data[["x_design_matrix"]]
Z = my_data[["z_design_matrix"]]

unif_lower_beta <- matrix(c(-5, -5, NA, NA), nrow = 2, byrow = TRUE)
unif_upper_beta <- matrix(c(5, 5, NA, NA), nrow = 2, byrow = TRUE)

unif_lower_gamma <- array(data = c(-5, NA, -5, NA, -5, NA, -5, NA),
                          dim = c(2,2,2))
unif_upper_gamma <- array(data = c(5, NA, 5, NA, 5, NA, 5, NA),
                          dim = c(2,2,2))

beta_prior_parameters <- list(lower = unif_lower_beta, upper = unif_upper_beta)
gamma_prior_parameters <- list(lower = unif_lower_gamma, upper = unif_upper_gamma)

modelstring = model_picker(prior = "uniform")
temp_model_file = tempfile()
tmps = file(temp_model_file, "w")
cat(modelstring, file = tmps)
close(tmps)

jags_model_object <- jags_picker(prior = "uniform",
                                 sample_size = n,
                                 dim_x = ncol(X), dim_z = ncol(Z),
                                 n_cat = 2,
                                 Ystar = obs_Y, X = X, Z = Z,
                                 beta_prior_parameters = beta_prior_parameters,
                                 gamma_prior_parameters = gamma_prior_parameters,
                                 number_MCMC_chains = 1,
                                 model_file = temp_model_file)
}