Compute Conditional Probability of Each Observed Outcome Given Each True Outcome, for Every Subject

Compute Conditional Probability of Each Observed Outcome Given Each True Outcome, for Every Subject

Usage

pistar_compute(gamma, Z, n, n_cat)

Arguments

gamma

A numeric matrix of regression parameters for the observed outcome mechanism, Y* | Y (observed outcome, given the true outcome) ~ Z (misclassification predictor matrix). Rows of the matrix correspond to parameters for the Y* = 1 observed outcome, with the dimensions of Z. Columns of the matrix correspond to the true outcome categories \(j = 1, \dots,\) n_cat.

Z

A numeric design matrix.

n

An integer value specifying the number of observations in the sample. This value should be equal to the number of rows of the design matrix, Z.

n_cat

The number of categorical values that the true outcome, Y, and the observed outcome, Y* can take.

Value

pistar_compute returns a matrix of conditional probabilities, \(P(Y_i^* = k | Y_i = j, Z_i) = \frac{\text{exp}\{\gamma_{kj0} + \gamma_{kjZ} Z_i\}}{1 + \text{exp}\{\gamma_{kj0} + \gamma_{kjZ} Z_i\}}\)

for each of the \(i = 1, \dots,\)

n subjects. Rows of the matrix correspond to each subject and observed outcome. Specifically, the probability for subject \(i\) and observed category $1$ occurs at row \(i\). The probability for subject \(i\) and observed category $2$ occurs at row \(i +\)

n. Columns of the matrix correspond to the true outcome categories \(j = 1, \dots,\)

n_cat.

Examples

if (FALSE) {
set.seed(123)
n <- 100
ones <- rep(1, n)
z <- rnorm(n)
Z <- matrix(c(ones, z), nrow = n, byrow = FALSE)
gamma <- matrix(c(1, 2, 3, 4), nrow = 2, byrow = FALSE)
conditional_probabilities <- pistar_compute(gamma, Z, n, n_cat = 2)
head(conditional_probabilities)
}