M-Step Expected Log-Likelihood with respect to Delta

Objective function of the form: \(Q_{\delta} = \sum_{i = 1}^N \Bigl[\sum_{j = 1}^2 \sum_{k = 1}^2 \sum_{\ell = 1}^2 w_{ij} y^*_{ik} \tilde{y}_{i \ell} \text{log} \{ \tilde{\pi}_{i \ell kj} \}\Bigr]\). Used to obtain estimates of \(\delta\) parameters.

Usage

q_delta_f(
  delta_v,
  V,
  obs_Ystar_matrix,
  obs_Ytilde_matrix,
  w_mat,
  sample_size,
  n_cat
)

Arguments

delta_v

A numeric array of regression parameters for the second-stage observed outcome mechanism, \(\tilde{Y} | Y^*, Y\) (second-stage observed outcome, given the first-stage observed outcome and the true outcome) ~ V (misclassification predictor matrix). The \(\delta\) vector is obtained from the array form. In array form, the first dimension (matrix rows) of delta corresponds to parameters for the \(\tilde{Y} = 1\) second-stage observed outcome, with the dimensions of the V The second dimension (matrix columns) correspond to the first-stage observed outcome categories \(Y^* \in \{1, 2\}\). The third dimension of delta_start corresponds to to the true outcome categories \(Y \in \{1, 2\}\). The numeric vector \(\delta\) is obtained by concatenating the delta array, i.e. delta_v <- c(delta_array).

V

A numeric design matrix.

obs_Ystar_matrix

A numeric matrix of indicator variables (0, 1) for the observed outcome Y*. Rows of the matrix correspond to each subject. Columns of the matrix correspond to each observed outcome category. Each row should contain exactly one 0 entry and exactly one 1 entry.

obs_Ytilde_matrix

A numeric matrix of indicator variables (0, 1) for the observed outcome \(\tilde{Y}\). Rows of the matrix correspond to each subject. Columns of the matrix correspond to each observed outcome category. Each row should contain exactly one 0 entry and exactly one 1 entry.

w_mat

Matrix of E-step weights obtained from w_j_2stage.

sample_size

An integer value specifying the number of observations in the sample. This value should be equal to the number of rows of the design matrix, V.

n_cat

The number of categorical values that the true outcome, Y, and the observed outcomes can take.

Value

q_beta_f returns the negative value of the expected log-likelihood function, \(Q_{\delta} = \sum_{i = 1}^N \Bigl[\sum_{j = 1}^2 \sum_{k = 1}^2 \sum_{\ell = 1}^2 w_{ij} y^*_{ik} \tilde{y}_{i \ell} \text{log} \{ \tilde{\pi}_{i \ell kj} \}\Bigr]\), at the provided inputs.