Set up a Naive Logistic Regression jags.model Object for a Given Prior

Set up a Naive Logistic Regression jags.model Object for a Given Prior

Usage

naive_jags_picker(
  prior,
  sample_size,
  dim_x,
  n_cat,
  Ystar,
  X,
  beta_prior_parameters,
  number_MCMC_chains,
  naive_model_file,
  display_progress = TRUE
)

Arguments

prior

character string specifying the prior distribution for the naive \(\beta\) parameters. Options are "t", "uniform", "normal", or "dexp" (double Exponential, or Weibull).

sample_size

An integer value specifying the number of observations in the sample.

dim_x

An integer specifying the number of columns of the design matrix of the true outcome mechanism, X.

n_cat

An integer specifying the number of categorical values that the true outcome, Y, and the observed outcome, Y* can take.

Ystar

A numeric vector of indicator variables (1, 2) for the observed outcome Y*. The reference category is 2.

X

A numeric design matrix for the true outcome mechanism.

beta_prior_parameters

A numeric list of prior distribution parameters for the \(\beta\) terms. For prior distributions "t", "uniform", "normal", or "dexp", the first element of the list should contain a matrix of location, lower bound, mean, or shape parameters, respectively, for \(\beta\) terms. For prior distributions "t", "uniform", "normal", or "dexp", the second element of the list should contain a matrix of shape, upper bound, standard deviation, or scale parameters, respectively, for \(\beta\) terms. For prior distribution "t", the third element of the list should contain a matrix of the degrees of freedom for \(\beta\) terms. The third list element should be empty for all other prior distributions. All matrices in the list should have dimensions dim_x X n_cat, and all elements in the n_cat column should be set to NA.

number_MCMC_chains

An integer specifying the number of MCMC chains to compute.

naive_model_file

A .BUG file and used for MCMC estimation with rjags.

display_progress

A logical value specifying whether messages should be displayed during model compilation. The default is TRUE.

Value

naive_jags_picker returns a jags.model object for a naive logistic regression model predicting the potentially misclassified Y*

from the predictor matrix x. The object includes the specified prior distribution, model, number of chains, and data.