EM-Algorithm Estimation of the Binary Outcome Misclassification Model while Assuming Perfect Sensitivity

EM-Algorithm Estimation of the Binary Outcome Misclassification Model while Assuming Perfect Sensitivity

Usage

perfect_sensitivity_EM(
  Ystar,
  Z,
  X,
  start,
  beta0_fixed = NULL,
  weights = NULL,
  expected = TRUE,
  tolerance = 1e-07,
  max_em_iterations = 1500
)

Arguments

Ystar

A numeric vector of indicator variables (1, 0) for the observed outcome Y*. The reference category is 0.

Z

A numeric matrix of covariates in the true outcome mechanism. Z should not contain an intercept.

X

A numeric matrix of covariates in the observation mechanism. X should not contain an intercept.

start

Numeric vector of starting values for parameters in the true outcome mechanism (\(\theta\)) and the observation mechanism (\(\beta\)), respectively.

beta0_fixed

Optional numeric vector of values of the observation mechanism intercept to profile over. If a single value is entered, this corresponds to fixing the intercept at the specified value. The default is NULL.

weights

Optional vector of row-specific weights used for selection bias adjustment. The default is NULL.

expected

A logical value indicating whether or not to calculate the covariance matrix via the expected Fisher information matrix. The default is TRUE.

tolerance

A numeric value specifying when to stop estimation, based on the difference of subsequent log-likelihood estimates. The default is 1e-7.

max_em_iterations

An integer specifying the maximum number of iterations of the EM algorithm. The default is 1500.

Value

perfect_sensitivity_EM returns a list containing nine elements. The elements are detailed in ?SAMBA::obsloglikEM documentation. Code is adapted from the SAMBA::obsloglikEM function.

References

Beesley, L. and Mukherjee, B. (2020). Statistical inference for association studies using electronic health records: Handling both selection bias and outcome misclassification. Biometrics, 78, 214-226.

Examples

set.seed(123)
n <- 1000
n_cat <- 2

ones <- rep(1, n)
x <- rnorm(n)
X <- matrix(c(ones, x), nrow = n, byrow = FALSE)
z <- rgamma(n, shape = 1)
Z <- matrix(c(ones, z), nrow = n, byrow = FALSE)

beta <- matrix(c(1, -2), ncol = 1)
gamma <- matrix(c(5, 20, -.5, -1), nrow = 2, byrow = FALSE)

probabilities <- pi_compute(beta, X, n, n_cat = 2)
#> Error in pi_compute(beta, X, n, n_cat = 2): could not find function "pi_compute"
conditional_probabilities <- pistar_compute(gamma, Z, n, n_cat = 2)
#> Error in pistar_compute(gamma, Z, n, n_cat = 2): could not find function "pistar_compute"

true_Y <- rep(NA, n)
for(i in 1:n){
  true_Y[i] = which(rmultinom(1, 1, probabilities[i,]) == 1)
}
#> Error in rmultinom(1, 1, probabilities[i, ]): object 'probabilities' not found
pistar_matrix_labels <- rep(1:n_cat, each = n)
obs_Y <- rep(NA, n)
for(i in 1:n){
 true_j = true_Y[i]
 obs_Y[i] = which(rmultinom(1, 1,
                            conditional_probabilities[c(i, n + i), true_j]) == 1)
 }
#> Error in rmultinom(1, 1, conditional_probabilities[c(i, n + i), true_j]): object 'conditional_probabilities' not found

obs_Y_reps <- matrix(rep(obs_Y, n_cat), nrow = n, byrow = FALSE)
category_matrix <- matrix(rep(1:n_cat, each = n), nrow = n, byrow = FALSE)
obs_Y_matrix <- 1 * (obs_Y_reps == category_matrix)

table(obs_Y, true_Y)
#> < table of extent 0 x 0 >

Ystar = ifelse(obs_Y == 1, 1, 0)
x_matrix = matrix(x, ncol = 1)
z_matrix = matrix(z, ncol = 1)

starting_values <- rep(1,6)
beta_start <- matrix(starting_values[1:2], ncol = 1)
gamma_start <- matrix(starting_values[3:4], ncol = 1, nrow = 2, byrow = FALSE)

perfect_sensitivity_results <- perfect_sensitivity_EM(Ystar,
                                                      Z = x_matrix, X = z_matrix,
                                                      start = c(beta_start, gamma_start),
                                                      beta0_fixed = NULL)
#> Error in perfect_sensitivity_EM(Ystar, Z = x_matrix, X = z_matrix, start = c(beta_start,     gamma_start), beta0_fixed = NULL): could not find function "perfect_sensitivity_EM"
perfect_sensitivity_results$param
#> Error in eval(expr, envir, enclos): object 'perfect_sensitivity_results' not found
c(beta, gamma[,2])
#> [1]  1.0 -2.0 -0.5 -1.0